+++ src/hooks/economics/useGetEgldPrice.ts
@@ -1,0 +1,16 @@
+import useSwr from 'swr';
+
+import { getEconomics } from 'apiCalls';
+import { ECONOMICS_ENDPOINT } from 'apiCalls/endpoints';
+
+export function useGetEgldPrice() {
+  const fetcher = (url: string) =>
+    getEconomics(url).then((response) => response.data);
+
+  const { data, error } = useSwr(ECONOMICS_ENDPOINT, fetcher);
+
+  return {
+    price: data?.price,
+    error
+  };
+}